Get changes from main repo

+++ java/java-analysis-impl/src/com/intellij/codeInspection/unneededThrows/RedundantThrows.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2000-2009 JetBrains s.r.o.
+ * Copyright 2000-2014 JetBrains s.r.o.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -268,7 +268,9 @@ public class RedundantThrows extends GlobalJavaBatchInspectionTool {
         if (!FileModificationService.getInstance().preparePsiElementsForWrite(refsToDelete)) return;
 
         for (final PsiJavaCodeReferenceElement aRefsToDelete : refsToDelete) {
-          aRefsToDelete.delete();
+          if (aRefsToDelete.isValid()) {
+            aRefsToDelete.delete();
+          }
         }
       }
       catch (IncorrectOperationException e) {
@@ -292,7 +294,10 @@ public class RedundantThrows extends GlobalJavaBatchInspectionTool {
 
       if (refMethod != null) {
         for (RefMethod refDerived : refMethod.getDerivedMethods()) {
-          removeException(refDerived, exceptionType, refsToDelete, (PsiMethod)refDerived.getElement());
+          PsiModifierListOwner method = refDerived.getElement();
+          if (method != null) {
+            removeException(refDerived, exceptionType, refsToDelete, (PsiMethod)method);
+          }
         }
       } else {
         final Query<Pair<PsiMethod,PsiMethod>> query = AllOverridingMethodsSearch.search(psiMethod.getContainingClass());
